Ph add authorization

+++ src/Controller/AnimalController.php
@@ -6,6 +6,7 @@ namespace App\Controller;
 use App\Entity\Animal;
 use App\Form\AnimalType;
 use App\Repository\AnimalRepository;
+use App\Security\AnimalVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,14 +15,22 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class AnimalController extends AbstractController
 {
+    private $animalRepository;
+
+    public function __construct(AnimalRepository $animalRepository)
+    {
+        $this->animalRepository = $animalRepository;
+    }
+
     /**
-     * @param AnimalRepository $animalRepository
      * @return Response
      * @Route("/", name="index")
      */
-    public function index(AnimalRepository $animalRepository): Response
+    public function index(): Response
     {
-        $animals = $animalRepository->findBy(['user' => $this->getUser()]);
+        $animals = $this->animalRepository->findBy(['user' => $this->getUser()]);
+        $this->denyAccessUnlessGranted(AnimalVoter::ACCESS, $animals);
+
         return $this->render(
             'animal/index.html.twig',
             [
@@ -71,6 +80,8 @@ class AnimalController extends AbstractController
      */
     public function show(Animal $animal): Response
     {
+        $this->denyAccessUnlessGranted(AnimalVoter::ACCESS, $animal);
+
         return $this->render(
             'animal/show.html.twig',
             [
@@ -86,6 +97,8 @@ class AnimalController extends AbstractController
      */
     public function delete(Animal $animal): Response
     {
+        $this->denyAccessUnlessGranted(AnimalVoter::ACCESS, $animal);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($animal);
         $em->flush();
@@ -107,6 +120,8 @@ class AnimalController extends AbstractController
      */
     public function update(Animal $animal, Request $request, EntityManagerInterface $em): Response
     {
+        $this->denyAccessUnlessGranted(AnimalVoter::ACCESS, $animal);
+
         $form = $this->createForm(AnimalType::class, $animal, ['btn-label' => "Aktualizuj"]);
         $form->handleRequest($request);
